ARM: dts: at91: sama5d2_som1_ek: disable ISC node by default
Git-commit: 85c1827eeee7fa8998fd57ec48b74fcc9ae9cf40
References: bsn#19
Patch-mainline: v5.10.76

[ Upstream commit 4348cc10da6377a86940beb20ad357933b8f91bb ]

Without a sensor node, the ISC will simply fail to probe, as the
corresponding port node is missing.
It is then logical to disable the node in the devicetree.
If we add a port with a connection to a sensor endpoint, ISC can be enabled.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>
Link: https://lore.kernel.org/r/20210902121358.503589-1-eugen.hristev@microchip.com
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.76-0005-ARM-dts-at91-sama5d2_som1_ek-disable-ISC-node-by-def.patch